requirements.txt
flask==3.0.2
tensorflow==2.15.0
pillow==10.2.0
numpy==1.26.4
gunicorn==21.2.0 